TYPROJECT/TY_Project/TY Project/backened/requirements.txt
flask
flask-sqlalchemy
flask-cors
python-dotenv
cloudinary
Pillow
colorthief
requests
google-generativeai
markdown
sqlalchemy
werkzeug
